Add file for proposal of @@clone / Reflect.clone()
I want this to be a thing some day maybe.

Maybe: there are probably things that should be done differently here, but it's not a bad start maybe? How about dealing with accessors vs data properties? How about asserting that the return type is of the same "kind" as the target of the cloning operation?